switched to for await

+++ src/commands/HousePoints.ts
@@ -1,6 +1,21 @@
-import { ActionRowBuilder, ButtonBuilder, ButtonStyle, ComponentType, EmbedBuilder, MessageActionRowComponentBuilder, PermissionFlagsBits, SlashCommandBuilder, TextChannel } from 'discord.js';
+import {
+    ActionRowBuilder,
+    ButtonBuilder,
+    ButtonStyle,
+    ComponentType,
+    EmbedBuilder,
+    MessageActionRowComponentBuilder,
+    PermissionFlagsBits,
+    SlashCommandBuilder,
+    TextChannel,
+} from 'discord.js';
 import { Client } from '../client/client.js';
-import { allPointChangeEmbed, LeaderboardButton, pointChangeButton, pointChangeEmbed } from '../util/builders.js';
+import {
+    allPointChangeEmbed,
+    LeaderboardButton,
+    pointChangeButton,
+    pointChangeEmbed,
+} from '../util/builders.js';
 import { Command, container } from '@sapphire/framework';
 import { ApplyOptions } from '@sapphire/decorators';
 import { House, ChannelId } from '../util/enum.js';
@@ -8,18 +23,20 @@ import { HousePoints } from '../database/DatabaseManager.js';
 
 @ApplyOptions<Command.Options>({
     name: 'housepoints',
-    description: 'Add or remove points from houses'
+    description: 'Add or remove points from houses',
 })
 export class HousePointsCommand extends Command {
     async chatInputRun(interaction: Command.ChatInputCommandInteraction) {
         const client = interaction.client as Client<true>;
-        const current = Object.fromEntries([...client.database.cache]) as HousePoints;
+        const current = Object.fromEntries([
+            ...client.database.cache,
+        ]) as HousePoints;
         const newTotals: HousePoints = {
             TIGER: interaction.options.getInteger('tigers') ?? current.TIGER,
             OWL: interaction.options.getInteger('owls') ?? current.OWL,
             RAVEN: interaction.options.getInteger('ravens') ?? current.RAVEN,
             TURTLE: interaction.options.getInteger('turtles') ?? current.TURTLE,
-            PANDA: interaction.options.getInteger('pandas') ?? current.PANDA
+            PANDA: interaction.options.getInteger('pandas') ?? current.PANDA,
         };
         const gifs = [
             'https://tenor.com/view/smiling-friends-adult-swim-smiling-friends-smiling-friends-spamish-pim-pimling-smiling-friends-pim-gif-10921009947340864978',
@@ -29,18 +46,28 @@ export class HousePointsCommand extends Command {
         ] as const;
 
         let changes = House.ids
-            .filter(house => newTotals[house] !== current[house])
-            .map(house => [house, newTotals[house]] as [House.id, number]);
+            .filter((house) => newTotals[house] !== current[house])
+            .map((house) => [house, newTotals[house]] as [House.id, number]);
 
         if (changes.length === 0)
-            return void interaction.reply({ content: 'No changes were made', ephemeral: true });
+            return void interaction.reply({
+                content: 'No changes were made',
+                ephemeral: true,
+            });
 
         const timeoutTimestamp = ~~((Date.now() + 60_000) / 1000);
 
         // TODO move changes to staged changes field
-        const stagedEmbed = allPointChangeEmbed(current, newTotals, interaction.user)
+        const stagedEmbed = allPointChangeEmbed(
+            current,
+            newTotals,
+            interaction.user
+        )
             .setTitle('Staged changes')
-            .addFields({ name: ':stopwatch: Timeout', value: `<t:${timeoutTimestamp}:R>` });
+            .addFields({
+                name: ':stopwatch: Timeout',
+                value: `<t:${timeoutTimestamp}:R>`,
+            });
 
         const commitButton = new ButtonBuilder()
             .setCustomId('commit')
@@ -60,40 +87,56 @@ export class HousePointsCommand extends Command {
         const reply = await interaction.reply({
             embeds: [stagedEmbed],
             components: [
-                new ActionRowBuilder<MessageActionRowComponentBuilder>()
-                    .addComponents(commitButton, cancelButton, partyButton)
+                new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                    commitButton,
+                    cancelButton,
+                    partyButton
+                ),
             ],
         });
 
-        const collector = reply.createMessageComponentCollector({ componentType: ComponentType.Button, time: 60_000 });
+        const collector = reply.createMessageComponentCollector({
+            componentType: ComponentType.Button,
+            time: 60_000,
+        });
 
         collector.on('collect', async (button) => {
             if (button.customId === 'party')
-                return void button.reply({ content: gifs[~~(Math.random() * gifs.length)], ephemeral: true });
+                return void button.reply({
+                    content: gifs[~~(Math.random() * gifs.length)],
+                    ephemeral: true,
+                });
 
             if (button.user.id !== interaction.user.id)
-                return void button.reply({ content: 'You do not have permission to use this', ephemeral: true });
+                return void button.reply({
+                    content: 'You do not have permission to use this',
+                    ephemeral: true,
+                });
 
-            if (button.customId === 'cancel')
-                return void reply.delete();
+            if (button.customId === 'cancel') return void reply.delete();
 
-            if (button.customId !== 'commit')
-                return void reply.delete();
+            if (button.customId !== 'commit') return void reply.delete();
 
             collector.stop();
 
             await button.update({
                 components: [
-                    new ActionRowBuilder<MessageActionRowComponentBuilder>()
-                        .addComponents(commitButton.setDisabled(true), cancelButton.setDisabled(true), partyButton.setDisabled(true))
+                    new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                        commitButton.setDisabled(true),
+                        cancelButton.setDisabled(true),
+                        partyButton.setDisabled(true)
+                    ),
                 ],
                 embeds: [
                     new EmbedBuilder()
                         .setColor('#2B2D31')
                         .setTitle('Committing changes...')
                         .setDescription(`This won't take long`)
-                        .setAuthor({ name: interaction.user.username, iconURL: interaction.user.displayAvatarURL() })
-                ]
+                        .setAuthor({
+                            name: interaction.user.username,
+                            iconURL: interaction.user.displayAvatarURL(),
+                        }),
+                ],
             });
 
             let time = performance.now();
@@ -105,65 +148,114 @@ export class HousePointsCommand extends Command {
                 const errorEmbed = new EmbedBuilder()
                     .setColor('#2B2D31')
                     .setTitle('Error')
-                    .setAuthor({ name: interaction.client.user.username, iconURL: interaction.client.user.displayAvatarURL() })
-                    .setDescription('Failed to update database, please try again later');
+                    .setAuthor({
+                        name: interaction.client.user.username,
+                        iconURL: interaction.client.user.displayAvatarURL(),
+                    })
+                    .setDescription(
+                        'Failed to update database, please try again later'
+                    );
 
-                await button.editReply({ embeds: [errorEmbed], components: [] });
+                await button.editReply({
+                    embeds: [errorEmbed],
+                    components: [],
+                });
 
-                return setTimeout(() => button.deleteReply().catch(console.error), 30_000);
+                return setTimeout(
+                    () => button.deleteReply().catch(console.error),
+                    30_000
+                );
             }
             time = performance.now() - time;
-    
+
             const embed = new EmbedBuilder()
                 .setColor(`#2B2D31`)
                 .setTitle('Changes pushed')
-                .setAuthor({ name: interaction.user.username, iconURL: interaction.user.displayAvatarURL() })
+                .setAuthor({
+                    name: interaction.user.username,
+                    iconURL: interaction.user.displayAvatarURL(),
+                })
                 .addFields(
-                    { name: 'Database patch time', value: `${time.toFixed(2)}ms` },
-                    { name: 'Changes', value: allPointChangeEmbed(current, newTotals, interaction.user).data.description ?? 'Error' }
+                    {
+                        name: 'Database patch time',
+                        value: `${time.toFixed(2)}ms`,
+                    },
+                    {
+                        name: 'Changes',
+                        value:
+                            allPointChangeEmbed(
+                                current,
+                                newTotals,
+                                interaction.user
+                            ).data.description ?? 'Error',
+                    }
                 );
-    
-            await button.editReply({ embeds: [embed], allowedMentions: { parse: [] }, components: [] });
-    
+
+            await button.editReply({
+                embeds: [embed],
+                allowedMentions: { parse: [] },
+                components: [],
+            });
+
             for (const houseId of House.ids) {
-                if (newTotals[houseId] === current[houseId])
-                    continue;
-    
+                if (newTotals[houseId] === current[houseId]) continue;
+
                 const changeButton = pointChangeButton(current, newTotals);
-                const actionRow = new ActionRowBuilder<MessageActionRowComponentBuilder>();
-                const embed = pointChangeEmbed(houseId, current[houseId], newTotals[houseId], interaction.user);
+                const actionRow =
+                    new ActionRowBuilder<MessageActionRowComponentBuilder>();
+                const embed = pointChangeEmbed(
+                    houseId,
+                    current[houseId],
+                    newTotals[houseId],
+                    interaction.user
+                );
                 const house = House[houseId];
-    
+
                 if (changeButton)
                     actionRow.addComponents(changeButton, LeaderboardButton());
-                else
-                    actionRow.addComponents(LeaderboardButton());
-    
+                else actionRow.addComponents(LeaderboardButton());
+
                 try {
-                    const channel = await client.channels.fetch(house.channelId) as TextChannel;
-    
+                    const channel = (await client.channels.fetch(
+                        house.channelId
+                    )) as TextChannel;
+
                     await channel.send({
                         embeds: [embed],
-                        components: [actionRow]
+                        components: [actionRow],
                     });
                 } catch (cause) {
-                    console.error(Error('Failed to send point changes to house channel', { cause }));
+                    console.error(
+                        Error('Failed to send point changes to house channel', {
+                            cause,
+                        })
+                    );
                 }
             }
-    
+
             try {
-                const [logs, trophy] = await Promise.all([client.channels.fetch(ChannelId.Logs), client.channels.fetch(ChannelId.Trophy)]) as [TextChannel, TextChannel];
-    
-                const payload = {
-                    embeds: [allPointChangeEmbed(current, newTotals, interaction.user)],
-                    allowedMentions: { parse: [] },
-                    components: [
-                        new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(LeaderboardButton())
-                    ]
-                };
-    
-                logs.send(payload);
-                trophy.send(payload);
+                for await (const channel of [
+                    client.channels.fetch(ChannelId.Logs),
+                    client.channels.fetch(ChannelId.Trophy),
+                ]) {
+                    if (!channel?.isTextBased()) continue;
+
+                    channel.send({
+                        embeds: [
+                            allPointChangeEmbed(
+                                current,
+                                newTotals,
+                                interaction.user
+                            ),
+                        ],
+                        allowedMentions: { parse: [] },
+                        components: [
+                            new ActionRowBuilder<MessageActionRowComponentBuilder>().addComponents(
+                                LeaderboardButton()
+                            ),
+                        ],
+                    });
+                }
             } catch (cause) {
                 console.error(Error('Failed to send logs', { cause }));
             }
@@ -171,8 +263,7 @@ export class HousePointsCommand extends Command {
 
         collector.on('end', (collected, reason) => {
             if (reason === 'time')
-                return void interaction.deleteReply()
-                    .catch(console.error);
+                return void interaction.deleteReply().catch(console.error);
         });
     }
 
@@ -183,18 +274,22 @@ export class HousePointsCommand extends Command {
             .setDefaultMemberPermissions(PermissionFlagsBits.ManageGuild);
 
         for (const house of House.ALL)
-            builder.addIntegerOption(
-                option => option
-                    .setName(house.id.toLowerCase().replace(/(\b\w+\b)/g, '$1s'))
+            builder.addIntegerOption((option) =>
+                option
+                    .setName(
+                        house.id.toLowerCase().replace(/(\b\w+\b)/g, '$1s')
+                    )
                     .setDescription(`New total for ${house.name}`)
             );
 
-        registry.registerChatInputCommand(builder, { guildIds: ['509135025560616963'] });
+        registry.registerChatInputCommand(builder, {
+            guildIds: ['509135025560616963'],
+        });
     }
 }
 
 container.stores.loadPiece({
     piece: HousePointsCommand,
     name: HousePointsCommand.name,
-    store: 'commands'
+    store: 'commands',
 });
